teste unitario lotes

+++ src/main/java/com/bootcamp_w3_g3/controller/ProdutoController.java
@@ -104,7 +104,7 @@ public class ProdutoController {
      * @return produtoDTO
      */
     @DeleteMapping(value="/deletar")
-    public void cadastro(@RequestParam Long cod_prod)
+    public void cadastro(@RequestParam Integer cod_prod)
     {
         produtoService.apagar(cod_prod);
     }
